+++ Promete.Example/examples/window/property.cs
@@ -18,13 +18,14 @@ public class property(ConsoleLayer console, Keyboard keyboard) : Scene
         console.Print($"UPS: {Window.UpdatePerSeconds}");
         console.Print($"FPS: {Window.FramePerSeconds}");
         console.Print($"Mode: {Window.Mode}");
-        // console.Print($"TotalFrame: {Window.TotalFrame}");
+
 
         console.Print($"[1]: Set WindowMode to {nameof(WindowMode.Resizable)}");
         console.Print($"[2]: Set WindowMode to {nameof(WindowMode.Fixed)}");
         console.Print($"[3]: Set WindowMode to {nameof(WindowMode.NoFrame)}");
         console.Print("[4]: Toggle Fullscreen: " + Window.IsFullScreen);
         console.Print("[5]: Toggle VSync: " + Window.IsVsyncMode);
+        console.Print("[6]: Toggle TopMost: " + Window.TopMost);
         console.Print("[ESC]: Exit");
 
         if (keyboard.Number1.IsKeyDown) Window.Mode = WindowMode.Resizable;
@@ -32,6 +33,7 @@ public class property(ConsoleLayer console, Keyboard keyboard) : Scene
         if (keyboard.Number3.IsKeyDown) Window.Mode = WindowMode.NoFrame;
         if (keyboard.Number4.IsKeyDown) Window.IsFullScreen = !Window.IsFullScreen;
         if (keyboard.Number5.IsKeyDown) Window.IsVsyncMode = !Window.IsVsyncMode;
+        if (keyboard.Number6.IsKeyDown) Window.TopMost = !Window.TopMost;
         if (keyboard.Escape.IsKeyDown) App.LoadScene<MainScene>();
     }
 }